Updated test case
